feat: Update image handling and source in index.astro
This commit improves the handling of images in the index.astro file by adding the Picture component and updating the usage of images throughout the file. It also updates the image source from "/assets/images/post-05.png" to "/assets/images/post-01.png", ensuring the accuracy of the displayed image on the page. These changes enhance the functionality and user experience of the page.